Use `type name(fields) is parent` as the declaration syntax
This makes it easier to see that the fields are associated with the type being declared, not with the parent type. It also makes it a bit more clear that the parent type does not get to know anything about the type being declared (and vice-versa).

The slight downside of this is that types with a larger amount of fields make it a bit more incovenient to see what's the parent type in the subtyping hierarchy. But this is a problem that's more easily solvable with IDE integrations.