Remove `just` as dependency in favour of Crochet's own tiny task runner (#13)
This reduces the amount of external trust Crochet needs to place on other things. The tiny task runner is trivial enough as it's only invoking a series of shell commands and defining a small dependency graph. just happens to include too many dependencies, one of which was tar, which had some open vulnerabilities. Crochet does not rely on tar in any way, and there's no point in bringing more code into Crochet than what it really needs.