A reasonable base for collections (#15)
This patch aims to take the current toy-level collection support in Crochet to a more practical (but not necessarily complete) collection support.

Core will have:

- Lists (trie-based persistent vectors)
- Maps (trie-based persistent maps)
- Sets (trie-based persistent sets)
- Linked lists (cons-based lists)
- Integer ranges (to reify 1 to: 10 by: 2)
- Streams (lazy, possibly procedural and/or infinite sequences)
- Records (records with extensible labels)

Along with the new types, this patch will introduce a fair amount of traits to allow functions to be shared among collection implementations more effectively, aid discoverability and reasoning, and better support future extension. Specialised implementations of the methods are likely going to be left to a separate patch.

Trie-based lists, linked lists, and integer ranges are going to be a separate patch. There's still a lot of work to be done on figuring streams and what role they should have in all of this as well...