Support global field declaration (#18)
By default all fields are local to the package in Crochet---that is, you can only project them inside of the package. This change allows one to explicitly declare a field as global, in which case we generate a command accessor for the field that can be used outside of the package.

The plan is to also do this for module and package local fields, but local commands are not supported yet.