Finer-grained capabilities through capability groups (#22)
Crochet uses types as capabilities, but initial capability grants are static, and described in linking constraints. This works nicely for more coarse grained capabilities, but ends up being too complicated to manage more fine-grained ones. We can see this even in the currently experimental standard library.

For example, `random` provides a means of having a dynamically global (global within a particular dynamic stack scope) random instance by defining a function that wraps a nullary block with an effect handler for providing the random instance. But this means that if `f` calls `g` and `g` calls `h`, then it's entirely possible that `g` could override this global instance for `h`---a powerful action that we might not want to allow.

Sadly, the only way of addressing it currently is moving the handler definition to a separate package, because capability grants use a strict package boundary. This is what was done with the `wall-clock` type, for example, which _should_ have been a part of the `crochet.time` package.

The additional intricacies of figuring out how to communicate between these packages, and how to maintain these separate packages, and how to distribute these separate packages, and---well, and so on. All of these tiny details will mean that people will almost certainly not care about separating powerful objects from static packages, and Crochet's security guarantees will suffer as a result.

This PR aims to address these tensions with the idea of "capability groups", which lets us have finer-grained static capabilities with a more reasonable management of them, and reasonable ways of reasoning about them.