Support for packaging the launcher with Electron (#28)
This allows the application to be alternatively distributed as a single zip with all one needs to start working with it. Distribution as a node package will still remain a supported option. Having things as a single zip simplifies the documentation and non-professional-programmers' lives immensely :)

The added command packages electron apps manually, does not do more than changing the executable name on the electron side of things, and does no signing. Ultimately this is a temporary thing, as the experience of using it is not exactly great. Whether one uses the node package or the electron one, both work by spawning a local HTTP server and connecting to it for the UI. The UI portion has no access to any electron capability.