Introduce trust hierarchies for text (#31)
This patch finally implements the trust hierarchies described in https://robotlolita.me/diary/2021/06/a-tale-of-4-strings/, and which have been on the to-do list for far too long. Libraries in stdlib that interact with the outside world now properly return untrusted-text for the bits of text that cannot be reasonably restricted to known semantics by Crochet. And the only thing one can do with untrusted-text is concatenate it elsewhere.

There'll need to be some more work on the UX of untrusted text in the future, as currently they're just a pain to use and reason about. The Core library now provides two capabilities in the mean time: `tainting` allows access to `taint make-untrusted: X`, which lets untrusted text be constructed from within Crochet code. And `untainting` allows access to `untaint make-trusted: X`, which lets one lift the taint mark in a piece of text, promoting it to a plain `text` type. Parsers currently do not make use of these directly, so a user needs to untaint before parsing---that's an absurd and insecure requirement that will be fixed before the stable release.

Also, more work needs to be done on how trust hierarchies interact with non-textual types. Particularly given how textual types move between structured representations (e.g.: JSON).