Introduce rich debug representations (#37)
This patch introduces a new debugging tool called "debug representations", where types can define multiple (rich and interactive) perspectives of how they'd like to be viewed in the playground, transcript, console, etc.

Representations and perspectives are capability-secure, and extensible without coordination (it's possible to add new perspectives and representations to an object without any other package being required to change its code or be aware of it).

This initial implementation uses a very restricted language for the representations because richer representations require far more security considerations (e.g.: think about the potential of phishing/confusing attacks if a package can simulate representations that look like native representations). So there will be follow ups to extend this language with the required security and privacy consideration that it requires.

This patch assumes code is primarily written and debugged from the interactive playground. Because the document language allows serialisation, this works even if objects live in different processes or programming languages, which is important for the Purr project.