Tracing for accurate, contextual time-travelling (#43)
This patch re-introduces the tracing mechanisms from earlier iterations of the VM, but using trace spans to keep track of the nesting contexts accurately. When applied to tracing commands, this allows for more useful traces of the program execution, since the way that the activation records nest is also reified in the traces---and this goes for things like effects and what not.

The features provided by this patch are very limited. They aim to provide a "minimal useful set of tracing tools", primarily for the A Crocheted Adventure book. Starting with a minimal set of tools makes it easier to iterate on the design, and does not require as much thought about all of the things that can go wrong when you let people have too much power over programs...

Hierarchical traces and tracing support for effects will be added in follow up PRs.