Ask for capability grants in the launcher (#46)
Whenever one opens a package, ask the user if they trust any new or modified capabilities for that name. The UI needs some improvements (particularly with grouping and better classifying risks, but that requires changes to the internal Crochet capability semantics first), and the description isn't currently lifted from the provider of the capability, which makes it difficult for the user to understand what exactly they're trusting---what are the implications of trusting it.

There is a related issue of grants being associated with package identifiers, but package identifiers not being guaranteed unique and there being no provenance guarantees (if you download a package with name something.goes.here, and you have previously trusted another package under that name, then this new package could just enjoy those capabilities without being asked for them). There are multiple steps to mitigate this, and the first one will be consolidating package imports (but not to be tackled in this PR).

So this is a start, but far from the ideal experience. Some serious work is needed here before this reaches a stable release.

This fixes #44