Rename `float` type so it is explicit about the size limitations (#50)
Non-size-suffixed type names should be reserved for arbitrary-precision types. If a type is bounded to a certain size, this should be reflected in its name. Doing so avoids confusing situations where people may not reason about the information bounds they're encoding in their programs, thus making programs unsafe.

So this patch renames the previous `float` type to `float-64bit`, and all native operations from `float` to some variation of `float-64`.

This fixes #40