Remove DSL (#54)
The idea of DSL was to better support embedded DSLs where users would have more control over how to interpret the syntax, but the syntax itself was fixed. It turned out to be cumbersome, and at the same time made it hard to provide good debugging tools for them, even with it being more static. Crochet will go down the path of having only external DSLs (properly integrated in the IDE, build system, and VM).

This fixes #53