Fix dispatch in the presence of placeholders (#57)
Historically, Crochet assumed that all declarations were ordered. This means that dispatch was resolved at the point new branches were added, which was fine then. But with placeholder types, this is now wrong. If you define a command over yet-to-be-defined types and then define a hierarchy later, the command's dispatch would not be resolved again.

Note that this patch does not solve the problem properly. It just adds calls to re-resolve dispatch once new types are introduced and a "more or less stable" shape can be assumed. Which, of course, breaks entirely in the playground/repl if you define placeholders. This will be fixed along with the proper optimisations for dispatch that are planned, but the current bug is a bit too pressing and annoying, so this is a bit of a hack to get that working and out of the way.