Intrinsic internal capability and package type (#61)
The idea is that the package type and internal capability can be the standard way of sealing types and commands that are meant to be internal. The package type also provides programmatic access to package internal data---currently this is used to implement sandboxed file systems for each package---in the browser this is done by generating a set of routes with opaque tokens in them, this token is then granted to each package and used to generate the URL for each asset file.

There's quite a lot that still needs to be done to make this actually safe, but won't be done in this PR. Plus this change does highlight the need of adding support for optional capabilities before the stable release, as otherwise we're forced to grant strong capabilities just to load a package.

This patch also removes the dependency on uuid, as the only one we care about (random 128 bit UUID) is provided natively by both browsers and Node.