A re-thought version of Novella (#80)
Version 2 of novella goes further in the direction of making hybrid parser-based IFs and VNs. This includes an auto-completion-based input for commands which ties in with Crochet's logic-based simulation. Largely a work in progress, not to be used for any serious project yet.

This is the same version used in https://github.com/qteatime/the-sound-of-rain