Add LJT (part 1/?) (#83)
This patch adds LJT, a data serialisation/deserialisation language which focuses on efficient space usage and cross-platform compatibility with binary data. The backends initially provided will be TypeScript, JavaScript, and Crochet, through very straight-forward template compilers.

In the future this will also support proper long-term storage (through versioning and migrations), data integrity checks (through contracts and integrity values), and extensible compilers (through annotations and backend plugins). As well as a proper IDE (there's a sketch of one in this patch, but things like time-travelling, annotating sections of the binary with the appropriate data labels, and so on require a bit more work).

LJT is a pre-requisite for building Purr and the stable Crochet VM.